requirements.txt
petl
flask
sqlalchemy
sqlalchemy-migrate
pluginbase
flask-login
flask-sqlalchemy
flask-security
flask-wtf